% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getxycoordinates.R
\name{getxycoordinates}
\alias{getxycoordinates}
\title{getxycoordinates}
\usage{
getxycoordinates(df)
}
\arguments{
\item{df}{The dataframe containing time and status data required to create the Kaplan Meier curve.}
}
\description{
This function converts a Kaplan Meier curve into xy coordinates. It returns a dataframe containing xy coordinates of the Kaplan Meier curve created with time-to-event data from the dataframe provided.
}
