% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMSubtractionAccuracy.R
\name{KMSubtractionAccuracy}
\alias{KMSubtractionAccuracy}
\title{KMSubtractionAccuracy}
\usage{
KMSubtractionAccuracy(
  n,
  mc = 1000,
  censor.p,
  subgroup.p,
  interval = 8,
  missing.p = 0,
  ncores = 1
)
}
\arguments{
\item{n}{Size of overall cohort.}

\item{mc}{Number of Monte Carlo iterations. Default is set at 1000.}

\item{censor.p}{Proportion of patients with censorship status, which may be found from the object of KMSubtractionMatch().}

\item{subgroup.p}{Proportion of subgroup.}

\item{interval}{Number of number-at-risk table intervals.}

\item{missing.p}{Proportion of missing data from the opposing subgroup. Default is set at 0.00.}

\item{ncores}{Number of cores to be utilized for parallel processing. The number of cores in your device may be found using the function 'parallel::detectCores()' from the parallel package.}
}
\value{
Imputed and original survival data were compared by means of marginal Cox-proportional hazard models and restricted mean survival time difference (RMSTD). This function returns density plots per matching algorithm for both ln(HR) and RMST-Difference; as well as summary statistics in table format.
}
\description{
This function conducts Monte Carlo simulations to evaluate the accuracy of KMSubtractionMatch() given parameters surrounding the imputation task required.
Follow-up time was modelled by a random weibull distribution of common shape parameter of 1.000 and scale parameter of 5.000.
}
\examples{
# Size of dataset, censorship proportion and subgroup proportion may be retrieved from the KMSubtractionMatch object, under Parameters.

data(cancer)
df_overall=colon
df_subgroup=colon[1:200,]
match=KMSubtractionMatch(df_overall, df_subgroup, matching="bipartite")

n=match$Parameters[1,1]
subgroup.p=match$Parameters[3,1]
censorship.p=match$Parameters[4,1]

KMSubtractionAccuracy(n=n,
mc=1000,
censor.p=censorship.p,
subgroup.p=subgroup.p,
interval=8,
missing.p=0.02,
ncores=5)
}
\keyword{}
