% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMSubtractionError.R
\name{KMSubtractionError}
\alias{KMSubtractionError}
\title{KMSubtractionError}
\usage{
KMSubtractionError(
  n,
  censor_overall.p,
  censor_subgroup.p,
  subgroup.p,
  interval = 8,
  missing.p = 0,
  mc = 1000,
  ncores = 1,
  fig.label = ""
)
}
\arguments{
\item{n}{Size of overall cohort.}

\item{censor_overall.p}{Proportion of patients with censorship status in the overall cohort, which may be found from the object of KMSubtractionMatch().}

\item{censor_subgroup.p}{Proportion of patients with censorship status in the subgroup cohort, which may be found from the object of KMSubtractionMatch().}

\item{subgroup.p}{Proportion of reported subgroup.}

\item{interval}{Number of number-at-risk table intervals.}

\item{missing.p}{Proportion of missing data from the opposing subgroup. Default is set at 0.00.}

\item{mc}{Number of Monte Carlo iterations. Default is set at 1000.}

\item{ncores}{Number of cores to be utilized for parallel processing. The number of cores in your device may be found using the function 'parallel::detectCores()' from the parallel package.}
}
\value{
Reconstructed and original survival data were compared by means of marginal Cox-proportional hazard models and restricted mean survival time difference (RMSTD). This function returns density plots per matching algorithm for both ln(HR) and RMST-Difference; as well as summary statistics in table format.
}
\description{
This function conducts Monte Carlo simulations to evaluate the limits of error of KMSubtractionMatch() given parameters surrounding the reconstruction task required.
Follow-up time was modeled by a random weibull distribution of common shape parameter of 1.000 and scale parameter of 5.000.
}
\examples{
Size of dataset, censorship proportion and subgroup proportion may be retrieved from the KMSubtractionMatch object, under Parameters.

data(cancer)
df_overall=colon
df_subgroup=colon[1:200,]
match=KMSubtractionMatch(df_overall, df_subgroup, matching="bipartite")

match$Parameters

n=match$Parameters[1,1]
subgroup.p=match$Parameters[3,1]
censor_overall.p=match$Parameters[4,1]
censor_subgroup.p=match$Parameters[5,1]

KMSubtractionError(n=n,
mc=1000,
censor_overall.p=censor_overall.p,
censor_subgroup.p=censor_subgroup.p,
subgroup.p=subgroup.p,
interval=8,
missing.p=0.02,
ncores=5)
}
\keyword{}
