% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMSubtractionEvaluateMatch.R
\name{KMSubtractionEvaluateMatch}
\alias{KMSubtractionEvaluateMatch}
\title{KMSubtractionEvaluateMatch}
\usage{
KMSubtractionEvaluateMatch(df_match)
}
\arguments{
\item{df_match}{A dataframe (data) from KMSubtractionMatch() containing time-to-event data of the matched cohorts "time" and "status"; the curve it originates from "curve"; as well as the "matchid".}
}
\description{
This function evaluates of the quality of matching by inspecting Empirical cumulative distribution functions and the Kolmogorov−Smirnov Test of follow-up times between matched pairs as well as with Bland-Altman plots to explore discrepancies between matched pairs.
}
\examples{
data(cancer)
df_overall=colon
df_subgroup=colon[1:200,]
match=KMSubtractionMatch(df_overall, df_subgroup, matching="bipartite")

KMSubtractionEvaluateMatch(df_match=match$data)
}
