% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMSubtractionMatch.R
\name{KMSubtractionMatch}
\alias{KMSubtractionMatch}
\title{KMSubtractionMatch}
\usage{
KMSubtractionMatch(df_overall, df_subgroup, matching = "bipartite")
}
\arguments{
\item{df_overall}{Dataframe of the overall cohort containing time-to-event data. Prepared as "time" and "status" respectively.}

\item{df_subgroup}{Dataframe of the subgroup cohort containing time-to-event data. Prepared as "time" and "status" respectively.}

\item{matching}{Choice of matching algorithm. The default matching algorithm is undertaken utilizing the minimum cost bipartite matching with the Hungarian algorithm ("bipartite"), users may choose to use Mahalanobis distance matching ("Maha") or nearest neighbor matching with distances determined by logistic regression ("logit").}
}
\description{
This function conducts matching of patients between subgroup and an overall cohort.
}
\examples{
data(cancer)
df_overall=colon
df_subgroup=colon[1:200,]
KMSubtractionMatch(df_overall, df_subgroup, matching="bipartite")
}
