% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMSubtractionMatch_Logit.R
\name{KMSubtractionMatch_Logit}
\alias{KMSubtractionMatch_Logit}
\title{KMSubtractionMatch_BP}
\usage{
KMSubtractionMatch_Logit(df_overall, df_subgroup)
}
\arguments{
\item{df_overall}{Dataframe of the overall cohort containing time-to-event data. Prepared as "time" and "status" respectively.}

\item{df_subgroup}{Dataframe of the subgroup cohort containing time-to-event data. Prepared as "time" and "status" respectively.}
}
\description{
This function conducts matching of patients between subgroup and an overall cohort through nearest neighbour matching with distances determined by logistic regression.
}
\examples{
KMSubtractionMatch_BP(df_overall, df_subgroup)
}
