% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getxycoordinates.R
\name{getxycoordinates}
\alias{getxycoordinates}
\title{Retrieve xy coordinates from time-to-event survival data.}
\usage{
getxycoordinates(df, label = 1)
}
\arguments{
\item{df}{The dataframe containing time and status data required to create the Kaplan Meier curve.}

\item{label}{Name the curve that will be exported. Useful in providing a unique id in parallel processing. Not needed in serial processing.}
}
\description{
This function converts a Kaplan Meier curve into xy coordinates. It returns a dataframe containing xy coordinates of the Kaplan Meier curve created with time-to-event data from the dataframe provided.
}
\examples{
data(cancer)
getxycoordinates(colon)
}
